fix(engine): very large ticks now accumulate bonus time
If the tick is larger than `MAXIMUM_TICK` (currently 2000ms), we set the tick
time to `MAXIMUM_TICK` and store the rest as bonus time. This can happen if the
machine goes into suspend with the game open.

This implementation means you can lose up to `MAXIMUM_TICK` msec if you suspend
right before a loop, but whatever.